+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/test/scala/com/azure/cosmos/spark/BulkWriterITest.scala
@@ -1096,7 +1096,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       // if the item does not exists, patchBulkUpdate essentially will create those items
       // Validate that patchBulkUpdate can create items successfully
       val writeConfig = CosmosWriteConfig(
-          ItemWriteStrategy.ItemPatchBulkUpdate,
+          ItemWriteStrategy.ItemBulkUpdate,
           5,
           bulkEnabled = true,
           bulkMaxPendingOperations = Some(900),
@@ -1130,7 +1130,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
 
       val writeConfig = CosmosWriteConfig(
-          ItemWriteStrategy.ItemPatchBulkUpdate,
+          ItemWriteStrategy.ItemBulkUpdate,
           5,
           bulkEnabled = true,
           bulkMaxPendingOperations = Some(900),